plugins: screensaver-proxy: Fix a segfault when uninhibiting
When an application requests that the screensaver be inhibited a cookie
is created for it and put into a hash table. The function
(csd-screensaver-proxy-manager.c::handle_method_call) that does this
also handles the uninhibiting case and would remove the cookie from the
hash table.

After this the name_vanished_cb callback function is called, it looks
through the cookie hash table for the relevant cookie and gets the
associated sender value. Unfortunately due to the above removal of the
cookie entry this never finds anything and then when it tries to use the
still uninitialised sender variable to remove an entry from the watch
hash table, you get a segfault.

Seeing as the above loop (in name_vanished_cb) removes the cookie entry
itself we can simply remove the call to g_hash_table_remove in
handle_method_call that removes the cookie entry.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>