plugins: screensaver-proxy: Allow /ScreenSaver interface path
This is essentially a port of Bastien Nocera's patch to
g-s-d/screensaver-proxy[0].

The problem was that under cinnamon 2.x, the current version of the
vlc[1] media player is using the /ScreenSaver path to
org.freedesktop.ScreenSaver which c-s-d screensaver-proxy didn't
recognise

    method call sender=:1.118 -> dest=org.freedesktop.ScreenSaver
     serial=3 path=/ScreenSaver; interface=org.freedesktop.ScreenSaver;
     member=Inhibit
       string "vlc"
       string "Playing some media."
    error sender=:1.24 -> dest=:1.118
     error_name=org.freedesktop.DBus.Error.UnknownMethod reply_serial=3
       string "No such interface 'org.freedesktop.ScreenSaver' on object
       at path /ScreenSaver"

and the screensaver would kick in after a time while watching a video.

This adds the definition for the /ScreenSaver freedesktop.org interface
path and allows vlc to properly inhibit the screensaver.

[0] -
https://git.gnome.org/browse/gnome-settings-daemon/commit/?h=gnome-3-10&id=fd0e2b107cdf5e82a034fc9a30753259ecc3d267
[1] -
http://git.videolan.org/?p=vlc.git;a=commitdiff;h=5fe4b004e254b6091f997b394190ad1609c67aa8

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>