power and media-keys: Add support for logind
Opt out of logind handling the sleep and power keys, and also inhibiting for lid close auto-suspend.

Also use a delay inihibit for logind so that we can do actions on suspend like blanking the screen using the screensaver and also poking the screensaver on resume.

This patch is based on the following gnome-settings-daemon patch, but keeps ConsoleKit and upower supported.
https://git.gnome.org/browse/gnome-settings-daemon/commit/?id=6defe42c31b18ad8dbb2fff3869b0fccbd821e97