Unown our DBus name when cinnamon-session says "Stop"
We don't want to quit yet because if we do, cinnamon and still
mapped windows lose their theme and icons. But we have to unown our
DBus name otherwise cinnamon-session will hang waiting for us.

This only works due to a bug in cinnamon-session where it handles any
client name being unowned as if the client has disconnected. Will need
to be revisited when that bug is fixed in cinnamon-session.

https://bugzilla.gnome.org/show_bug.cgi?id=727049