Xrandr plugin: Do not save configuration when RotateTo is called.
On systems equipped with accelerometers, the orientation plugin triggers dbus calls to "RotateTo" when the orientation of the screen changes.

This was saving the configuration and had the following effect:

 - The new rotation was successfully saved in .config/monitors.xml
 - The previous rotation was saved in .config/monitors.xml.backup

When a configuration is changed from say.. "System settings -> Display", the very same thing happens, but a dialog is shown to the user to confirm his/her choice. On confirmation, the backup file is then removed.

On session startup, the xrandr plugin in CSD checks for the presence of a backup file. If found, CSD assumes there had been a crash or the user had been unable to confirm... and consequently overwrites the configuration with the backup.

Here is a typical example illustrating why this is problematic:

 - The user rotates his laptop right, and then back to normal.
 - CSD puts a normal rotation in .config/monitors.xml and a right rotation in .config/monitors.xml.backup
 - No dialog is shown (and that's good) and the backup file isn't deleted (and that's bad)
 - On the next login, Cinnamon overwrites the configuration because of the presence of the backup file and thus Cinnamon starts with a right rotation.

This produces unwanted backup restorations and rotation glitches first.

Even worse, on selected hardware, where we observe kernel issues (this is the case in Mint 18), sensors do not necessarily work after a cold reboot. In this case, the screen doesn't even rotate back to normal and is stuck at 90 degrees.

There are two ways to fix this issues:

 - Ensure that the configuration is confirmed (we don't want a confirmation dialog here obviously, but we do want the backup file to be deleted and/or not created).
 - Apply the rotation without saving it into the configuration.

The latter solution was chosen in this commit.

Modifying the rotation using "System Settings -> Displays" still works and still has permanent effects (because it calls "ApplyConfiguration" as opposed to "RotateTo").

With this change, the orientation plugin simply rotates on the fly and no longer impacts the saved configuration.