Don't kill plugins during CSM stop PHASE
We want CSD plugins to run for as long as possible, until the very very
end of the session, because we don't want to lose xrandr, mouse support,
desktop themeing and whatnot.

Plugins respond to QueryEndSession and EndSession phases already.

CSM no longer waits for processes lagging during the Stop phase so this
commit should not introduce lag in the shutdown/logout sequence.