Don't set GDK_SCALE=1 in automount plugin.
When connecting an encrypted HDD over USB, the automont plugin
calls g_volume_mount on it. This results in showing a dialog window.
Because of GDK_SCALE=1 this dialog looks squashed (similar to a GTK2 window)
in HiDPI mode.

Note: GDK_SCALE=1 is need for the background plugin to work properly.

Note: Regressions might occur in other plugins if GDK_SCALE isn't set. While working
on this commit, regressions could be observed when playing with cursor themes and
fonts in HiDPI mode. To reduce the risk of regressions, this commit focuses on
only fixing the CSD-automount dialog and thus only sets the automount plugin to
FORCE_GDK_SCALE=FALSE.