automount: Add a .desktop file
Cinnamon doesn't like untitled windows and falls back to appsys.

This fixes the title and icon of the dialog popping up when plugging
in an encrypted volume.

It now uses the folder icon and "Files" as the title, in line with
what happens when the volume is mounted via the Nemo sidebar.