Xrandr: Hold the session start while the plugin is loaded.
This is an attempt to prevent artefacts from showing on the screen. These are
visible "sometimes" after login in, when using Cinnamon in dual miniDP+HDMI mode
with two high resolutions (2560x1440). Artefacts go away by launching the Display
settings and pressing the apply button.

Hopefully with this commit they won't happen at all anymore.

This should/could be reverted if the issue is observed again.

Note: That issue is visible during boot time (plymouth) but never seems to impact
the Slick Greeter.