CSD: Add menu to CSD windows when the layout is :minimize,maximize,close
This is a quick temporary fix.

CSD headerbars need menu, as their menu button usually contains features
which aren't anywhere else to be found within the apps.

In normal windows on the other hand, menu is completely useless as it
duplicates what's already in the titlebar context menu.

Ideally we should parse the layout and add menu when it's missing.

This is a quick fix which simply adds menu to the default layout.