housekeeping: fix improper notify_notification_close() usage
notify_notification_close() expects that a parameter will be available for
the error location, which could be a dangling pointer in a register or
on the stack in the case of some architectures.

This was caught by GNOME/glib#1425 which allows us to check proper type
parameters.

Cherry-picked to cinnamon-settings-daemon from
https://gitlab.gnome.org/GNOME/gnome-settings-daemon/commit/3110457f