csd-power-manager.c: Fix conversion of absolute keyboard brightness back to percent, and fix signal handling of the keyboard backlight proxy.
In the method handler, we were attempting to convert what was already
a percentage into percentage, which was causing warnings during runtime.

In the signal handler, there are two signals rather than one, each with
a different signature.  Listen to the one that provides a source name,
so we can keep from reacting to our own SetBrightness calls (like in the
method handler.)