Fix up the configure check for dbus directories.
In commit 139c7137525ab29ddfc4a6abfd6287cd3f626c22 the datetime plugin
was (re)added to the tree, but the checks designed to control where the
dbus files were placed, never did anything.

Make this work, and also change the default location of system.d to be
more consistent with the services directory, i.e. in the datadir which
is typically reserved for packaged software.

debian:
- update files whitelist to match the new dbus install dir
- install maint script to remove conffile if it is unmodified. Debian
  doesn't do this automatically.