csd-power-manager.c: Check for a device battery-level property before attempting to access it.
The BatteryLevel property was introduced in UPower 0.99.5.  On anything
older than that, a warning would be thrown, and garbage passed to any
dbus power clients (like cinnamon-settings).

Now, if the battery-level property doesn't exist, a safe 'not supported'
value will be returned, indicating  that the percentage should be used
instead.  We can continue to use the DeviceLevel values, since they are
shared with the warning-level property.