debian/cinnamon-settings-daemon.links: Fix typo.
/usr/bin/csd-a11y-keyboard is hardcoded in ubiquity-dm, and is also
specifically the file checked when determining whether it should assume
a cinnamon environment.  With this failing, no other csd- programs run,
including csd-xsettings, which sets the correct themes for the session.

This didn't affect the normal user session, as it's started by an autostart
desktop file, which calls the binary directly (/usr/lib/*) rather than
the /usr/bin link.