csd-media-keys: Check for the availability of a mouse/keyboard grab when running screenshot commands.
This introduces a small delay when running a command, and in the case
of gnome-screenshot, will retry a few times if it can't get a grab
immediately.

The delay applies to all commands, which helps prevent excessive windows
spawning if a key is held down too long (this is imperfect - it will
still allow a single additional window due to the delay being shorter
than the default keyboard repeat delay).