csd-power-manager.c: Use a different method to determine whether or not the system is on battery.
The up_client_on_battery() method doesn't update as it should, but
the device state seem to be reliable, so base it on that instead -
this was preventing low/critical notifications from being sent when
appropriate.

Also make sure to figure out the correct warning level at startup,
which will make sure a notification is sent after logging in (and
not just if the warning level changes during a session).