power: Increase the critical and action thresholds so they are higher than the upower default.
This is a stopgap - in an upcoming commit, we'll move to rely entirely
on upower's device levels (in /etc/UPower/UPower.conf) as well as let
it handle the critical action. Its interference is unavoidable, but
by default it behaves nearly the same as our power plugin.

By increasing our own threshold slightly, it ensures our action
preference gets first shot at controlling the critical action.