csd-wacom: remove ifdef code that cannot compile (#395)
Regression in commit 88e9de536db46ed1bda3777d24e9305d694099c6. This was
probably a bad merge from gnome. The HAVE_WAYLAND define is never
defined, but if it did then the code would try to include a cinnamon
header that has never existed, then call a gnome-settings-daemon
internal function that does not exist.

It is possible to properly implement HAVE_WAYLAND, but this code didn't.