Revert "csd-media-keys: Show device description on volume OSD (#399)"
This reverts commit 1dbe86a36801f2bb12284f4c10e0bd978bb7178c.

It should be obvious to the user which output device is active and
interacted with. We only have one set of media keys on the computer.
This also always show the device name, even when the computer only
has one.

This would be useful if we had multiple output devices and multiple
sets of media keys (like on a sound mixer for instance), but it's not
the case for most people.